chore(wallet-core): extract entropy check thunks
